{"schema":"decodex/commit/1","summary":"Move scroll capture sample state","authority":"manual"}
